fix: resolve all P0 safety bugs — format-string crash, int overflow, reparse-point gaps, cloud-file detection, TOCTOU
- Fix format-string crash in MinFolderAge skip logging (missing args to string.Format)
- Change file size from int to long to prevent truncation of files >2GB
- Add reparse-point check on start folder (junctions/symlinks/mount points refused)
- Add defense-in-depth reparse guard in SecureDeleteDirectory
- Detect OneDrive/cloud-only placeholder files as real content (FILE_ATTRIBUTE_RECALL_ON_DATA_ACCESS)
- Replace TOCTOU-vulnerable IsDirLocked rename-test with ACL-based permission check
- Fix pre-build event to use PowerShell instead of hardcoded author-specific date.exe path
- Retarget to .NET Framework 4.8.1 (superset of 4.8, targeting pack available)

+++ RED/RedLib/DeletionWorker.cs
@@ -138,7 +138,7 @@ namespace RED
 					FileInfo file = Files[f];
 
 					string delPattern;
-					bool deleteTrashFile = this.Data.IgnoreFileNameList.IsOnList(file, (int)file.Length, this.Data.IgnoreEmptyFiles, out delPattern);
+					bool deleteTrashFile = this.Data.IgnoreFileNameList.IsOnList(file, file.Length, this.Data.IgnoreEmptyFiles, out delPattern);
 
 					// If only one file is good, then stop.
 					if (deleteTrashFile)
